pyproject.toml
[project]
name = "app-z-to-dht"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pycryptodome>=3.23.0",
]

[tool.ruff.format]
docstring-code-format = true
